index projects in the background in an actor
